+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # id3 Finger SDK Samples Changelog
 
+## Version 3.0.1.0 (2023-01-12)
+- Updated to id3 Finger SDK 3.0.1
+- Added java samples:
+    - RecognitionCLI: simple extraction and comparison process from images
+    - RecognitionNIST: extraction process and comparisons between two ANSI/NIST-ITL records
+
 ## Version 3.0.0.0 (2022-10-05)
 - Initial release for id3 Finger SDK 3.0.0